Designated projects as tests instead of class libraries
The project is recognized as a test project when the respective GUIDs is provided.
Also, a flask icon is added to a project's icon in Visual Studio's Solution Explorer when the project is designated as test.
The GUID infos are acquired from codeproject.com/reference/720512/list-of-visual-studio-project-type-guids and default VS's default test project templates.